Server/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
alembic==1.14.0
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
apispec==6.7.1
appdirs==1.4.4
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
audioread==3.0.1
auth0-python==4.7.2
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bidict==0.23.1
black==24.10.0
blinker==1.9.0
blis==0.7.11
build==1.2.2.post1
CacheControl==0.14.1
cachetools==5.5.0
catalogue==1.0.2
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.20
cleo==2.1.0
click==8.1.7
cloudpathlib==0.20.0
cohere==5.11.4
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
crashtest==0.4.1
crewai==0.80.0
crewai-tools==0.14.0
cryptography==43.0.3
cycler==0.12.1
cymem==2.0.10
dataclasses-json==0.6.7
debugpy==1.8.9
decorator==5.1.1
Deprecated==1.2.15
deprecation==2.1.0
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docstring_parser==0.16
docx2txt==0.8
dulwich==0.21.7
durationpy==0.9
EditorConfig==0.12.4
embedchain==0.1.125
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz#sha256=06c80936324012d1223291d2af41a5229e746dc2dee8fe31a532666ee3d18aaa
entrypoints==0.4
executing==2.1.0
faiss-cpu==1.7.4
Faker==33.0.0
fastapi==0.115.5
fastavro==1.9.7
fastjsonschema==2.20.0
favicon==0.7.0
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-PyMongo==2.3.0
flask-smorest==0.45.0
Flask-SocketIO==5.5.0
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
gast==0.6.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-aiplatform==1.73.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gptcache==0.1.44
groq==0.12.0
grpc-google-iam-v1==0.13.1
grpcio==1.68.0
grpcio-status==1.68.0
grpcio-tools==1.68.0
h11==0.14.0
h2==4.1.0
h5py==3.12.1
hpack==4.0.0
htbuilder==0.7.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
imageio==2.36.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
installer==0.7.0
instructor==1.6.4
ipykernel==6.29.5
ipython==8.29.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
jobspy==0.29.0
jsbeautifier==1.15.1
json_repair==0.30.2
jsonpatch==1.33
jsonpickle==4.0.0
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==2.6.0
Keras-Preprocessing==1.1.2
keyring==24.3.1
kiwisolver==1.4.7
kubernetes==31.0.0
lancedb==0.16.0
langchain==0.3.7
langchain-cohere==0.3.2
langchain-community==0.3.7
langchain-core==0.3.20
langchain-experimental==0.3.3
langchain-groq==0.2.1
langchain-huggingface==0.1.2
langchain-openai==0.2.9
langchain-text-splitters==0.3.2
langcodes==3.5.0
langsmith==0.1.145
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.9.2
lime==0.2.0.1
litellm==1.52.14
llvmlite==0.43.0
lxml==5.3.0
Mako==1.3.6
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mem0ai==0.1.32
ml-dtypes==0.4.1
mmh3==5.0.1
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.11
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.14.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
nodeenv==1.9.1
numba==0.60.0
numpy==1.26.0
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.55.0
opencv-python==4.10.0.84
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.11
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.2
parameterized==0.9.0
parso==0.8.4
pathspec==0.12.1
pdfminer==20191125
pdfminer.six==20240706
pdfminer3==2018.12.3.0
pexpect==4.9.0
pillow==10.4.0
pkginfo==1.11.2
plac==1.1.3
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
pooch==1.8.2
portalocker==2.10.1
posthog==3.7.2
preshed==3.0.9
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.1
pydantic-settings==2.6.1
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.10.0
pylance==0.19.2
pymdown-extensions==10.12
pymongo==4.10.1
PyMySQL==1.1.1
pyparsing==3.2.0
pypdf==5.0.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyresparser==1.0.6
pyright==1.1.389
pyrsistent==0.20.0
pysbd==0.3.4
PySocks==1.7.1
pysqlite3==0.5.4
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.1
python-socketio==5.12.0
pytube==15.0.0
pytz==2024.2
pyvis==0.3.2
PyYAML==6.0.2
pyzmq==26.2.0
qdrant-client==1.12.1
RapidFuzz==3.10.1
ratelim==0.1.6
redis==5.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
resampy==0.4.3
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
safetensors==0.4.5
schema==0.7.7
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.26.1
sentence-transformers==3.3.1
sentencepiece==0.2.0
shapely==2.0.6
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.6
spacy==2.3.9
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.35
srsly==1.0.7
st-annotated-text==4.0.1
st-theme==1.2.3
stack-data==0.6.3
starlette==0.41.3
streamlit==1.40.1
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.5.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-modal==0.1.2
streamlit-option-menu==0.4.0
streamlit-tags==1.2.8
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
thinc==7.4.6
threadpoolctl==3.5.0
tifffile==2024.9.20
tiktoken==0.7.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.1.0
tomli_w==1.1.0
tomlkit==0.13.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.0
traitlets==5.14.3
transformers==4.47.0
trio==0.27.0
trio-websocket==0.11.1
trove-classifiers==2024.10.21.16
typer==0.13.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uv==0.5.4
uvicorn==0.32.1
uvloop==0.21.0
validators==0.34.0
virtualenv==20.28.0
wasabi==0.10.1
watchdog==6.0.0
watchfiles==0.24.0
wcwidth==0.2.13
weasel==0.4.1
webargs==8.6.0
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
wrapt==1.17.0
wsproto==1.2.0
xattr==1.1.0
yarl==1.18.0
zipp==3.21.0
